backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.8.1
email-validator==2.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
slowapi==0.1.9
python-json-logger==3.2.1
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1
aiosqlite==0.20.0
ruff==0.9.7
